Add HTTP request timeout

+++ src/main/java/com/amplitude/Constants.java
@@ -3,7 +3,7 @@ package com.amplitude;
 public interface Constants {
 
     String  API_URL                 = "https://api2.amplitude.com/2/httpapi";
-    long    NETWORK_TIMEOUT_MILLIS  = 10000;
+    int     NETWORK_TIMEOUT_MILLIS  = 10000;
     String  SDK_LIBRARY             = "amplitude-java";
     String  SDK_VERSION             = "1.1.0";
